Move spec path resolution into index layer
